型定義を一箇所にまとめた
