+++ front/app/_components/AddPersonDialog.tsx
@@ -2,42 +2,7 @@ import React from 'react';
 import Dialog from '@mui/material/Dialog';
 import { useState } from "react";
 import { useSession } from 'next-auth/react';
-
-interface Member {
-  id: number;
-  name: string;
-  singles_total_game: number;
-  singles_win_game: number;
-  singles_strength: number;
-  doubles_total_game: number;
-  doubles_win_game: number;
-  doubles_strength: number;
-  history: (SinglesRecord | DoublesRecord)[];
-}
-
-interface SinglesRecord {
-  id: number;
-  player_1: string;
-  score_1: number;
-  player_2: string;
-  score_2: number;
-  user_id: number;
-  created_at: Date;
-  updated_at: Date;
-}
-
-interface DoublesRecord {
-  id: number;
-  player_1: string;
-  player_2: string;
-  score_12: number;
-  player_3: string;
-  player_4: string;
-  score_34: number;
-  user_id: number;
-  created_at: Date;
-  updated_at: Date;
-}
+import type { Member } from '@/app/types/index';
 
 export default function DialogSelect({ addPersonOpen, handleAddPersonClose, fetchData }: any) {
   const [member, setMember] = useState({} as Member);
