+++ front/app/_components/Members.tsx
@@ -14,42 +14,7 @@ import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
 import Button from '@mui/material/Button';
 import DeleteIcon from '@mui/icons-material/Delete';
-
-interface Member {
-  id: number;
-  name: string;
-  singles_total_game: number;
-  singles_win_game: number;
-  singles_strength: number;
-  doubles_total_game: number;
-  doubles_win_game: number;
-  doubles_strength: number;
-  history: (SinglesRecord | DoublesRecord)[];
-}
-
-interface SinglesRecord {
-  id: number;
-  player_1: string;
-  score_1: number;
-  player_2: string;
-  score_2: number;
-  user_id: number;
-  created_at: Date;
-  updated_at: Date;
-}
-
-interface DoublesRecord {
-  id: number;
-  player_1: string;
-  player_2: string;
-  score_12: number;
-  player_3: string;
-  player_4: string;
-  score_34: number;
-  user_id: number;
-  created_at: Date;
-  updated_at: Date;
-}
+import type { Member } from '@/app/types/index';
 
 function Row({ member, handleDelete }: {
                member: Member; handleDelete: (id: number) => void; }) {
@@ -139,7 +104,7 @@ function Row({ member, handleDelete }: {
   );
 }
 
-export default function Member({members, handleDelete}: 
+export default function Members({members, handleDelete}: 
                                {members: Member[]; handleDelete: (id: number) => void; }) {
 
   return (
