+++ front/app/page.tsx
@@ -2,44 +2,9 @@
 
 import { useCallback, useEffect, useState } from "react";
 import { useSession } from 'next-auth/react';
-import Member from './_components/Member';
+import Members from './_components/Members';
 import SpeedDialTooltipOpen from './_components/SpeedDialTooltipOpen';
-
-interface Member {
-  id: number;
-  name: string;
-  singles_total_game: number;
-  singles_win_game: number;
-  singles_strength: number;
-  doubles_total_game: number;
-  doubles_win_game: number;
-  doubles_strength: number;
-  history: (SinglesRecord | DoublesRecord)[];
-}
-
-interface SinglesRecord {
-  id: number;
-  player_1: string;
-  score_1: number;
-  player_2: string;
-  score_2: number;
-  user_id: number;
-  created_at: Date;
-  updated_at: Date;
-}
-
-interface DoublesRecord {
-  id: number;
-  player_1: string;
-  player_2: string;
-  score_12: number;
-  player_3: string;
-  player_4: string;
-  score_34: number;
-  user_id: number;
-  created_at: Date;
-  updated_at: Date;
-}
+import type { Member } from '@/app/types/index';
 
 export default function Home() {
   const [members, setMembers] = useState([] as Member[]);
@@ -77,7 +42,7 @@ export default function Home() {
     <>
       {session ?
         <>
-          <Member
+          <Members
             members={members}
             handleDelete={handleDelete}
           />
